% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cjmle_expr_simu}
\alias{cjmle_expr_simu}
\title{Simulation for the book chapter}
\usage{
cjmle_expr_simu(response, nonmis_ind, theta0, A0, A_true, cc, tol,
  max_steps = 1000L, print_proc = FALSE)
}
\description{
Simulation for the book chapter
}
