% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sim}
\alias{data_sim}
\title{Simulated dataset for item factor analysis on the multidimensional two parameter logistic model.}
\format{An object of class \code{list} of length 9.}
\usage{
data_sim
}
\description{
The dataset contains the simulation setting and the response data.
}
\keyword{datasets}
