% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getMIRTthreads}
\alias{getMIRTthreads}
\title{Get the number of threads that mirtjml is using}
\usage{
getMIRTthreads()
}
\value{
The number of threads that mirtjml is using
}
\description{
Get the number of threads that mirtjml is using
}
