% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirtjml.R
\name{mirtjml}
\alias{mirtjml}
\title{Exploratory item factor analysis for high dimensinal dichotomous responses data}
\usage{
mirtjml(response, nonmis_ind, K, theta0 = NULL, A0 = NULL, d0 = NULL,
  C = NULL, tol = 1e-04, trans_ogive = TRUE, print_proc = TRUE)
}
\arguments{
\item{response}{Matrix contains 0/1 responses for N rows of examinees and J columns of items.}

\item{nonmis_ind}{Matrix contains 0/1 elements with 1 stands for non-missing response in the coresponding position.}

\item{K}{Integer for specifying the latent dimension in exploratory item factor analysis.}

\item{theta0}{The initial value for person parameters. The default is set to be null and a 
default value based on singular value decomposition of resposne data will be computated.}

\item{A0}{The initial value for item parameters. The default is set to be null and a 
default value based on singular value decomposition of resposne data will be computated.}

\item{d0}{The initial value of the intercept parameters. The default is set to be null and a 
default value based on singular value decomposition of resposne data will be computated.}

\item{C}{The constraint constant. The default is set be to null and a default value is computed 
based on the given latent dimension K.}

\item{tol}{The tolerance parameter. The default value is 1e-4.}

\item{trans_ogive}{Is the result loading parameters transformed to ogive model, default is TRUE.}

\item{print_proc}{Print the precision during the esitmation procedure. The default is TRUE}
}
\value{
The function returns a list with the following components:
\describe{
  \item{theta_hat}{The estimated person parameters matrix.}
  \item{A_hat}{The estimated loading parameters matrix}
  \item{d_hat}{The estimated value of intercept parameters.}
}
}
\description{
Exploratory item factor analysis for high dimensinal dichotomous responses data
}
\references{
Joint Maximum Likelihood Estimation for High-dimensional Exploratory Item Factor Analysis
}
