% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setMIRTthreads}
\alias{setMIRTthreads}
\title{Set the number of threads that mirtjml should use}
\usage{
setMIRTthreads(threads = NULL)
}
\arguments{
\item{threads}{NULL (default) rereads environment variables. 0 means to use all logical CPUs available. Otherwise a number >= 1}
}
\value{
The number of threads that mirtjml was using previously
}
\description{
Set the number of threads that mirtjml should use
}
